// riskaware.client/src/components/enums/RiskAttributesEnum.tsx
export enum Probability {
  Insignificant = 1,
  Minor = 2,
  Moderate = 3,
  Major = 4,
  Severe = 5
}

export enum Impact {
  Insignificant = 1,
  Minor = 2,
  Perceptible = 3,
  Critical = 4,
  Catastrophic = 5
}

export enum Prevention {
  Neglect = "Zanedbat",
  Insurance = "Pojistit",
  Treatment = "Ošetřit"
}

export enum Status {
  Concept = "Koncept",
  Active = "Aktivní",
  Closed = "Uzavřené",
  Incident = "Přihodilo se"
}

export enum Category {
  New = -1,
  Finnancial = 2,
  Personnel = 3,
  Operational = 4,
  Legal = 5,
  Technical = 6
}